refactor(tests): cleanup and better separate protractor's basic test suite
