+++ spec/basic/lib_spec.js
@@ -1,5 +1,4 @@
 var util = require('util');
-var port =  process.env.HTTP_PORT || '8000';
 
 describe('no protractor at all', function() {
   it('should still do normal tests', function() {
@@ -17,10 +16,11 @@ describe('protractor library', function() {
     expect($).toBeDefined();
   });
 
-  it('should wrap webdriver', function() {
-    browser.get('index.html');
-    expect(browser.getTitle()).toEqual('My AngularJS App');
-  });
+  it('should export other webdriver classes onto the global protractor',
+      function() {
+        expect(protractor.ActionSequence).toBeDefined();
+        expect(protractor.Key.RETURN).toEqual('\uE006');
+      });
 
   it('should export custom parameters to the protractor instance', function() {
     expect(browser.params.login).toBeDefined();
@@ -29,25 +29,16 @@ describe('protractor library', function() {
   });
 
   it('should allow a mix of using protractor and using the driver directly',
-    function() {
-      browser.get('index.html');
-      expect(browser.getCurrentUrl()).
-          toEqual('http://localhost:'+port+'/index.html#/form');
-
-      browser.driver.findElement(protractor.By.linkText('repeater')).click();
-      expect(browser.driver.getCurrentUrl()).
-          toEqual('http://localhost:'+port+'/index.html#/repeater');
-
-      browser.navigate().back();
-      expect(browser.driver.getCurrentUrl()).
-          toEqual('http://localhost:'+port+'/index.html#/form');
-    });
-
-  it('should export other webdriver classes onto the global protractor',
       function() {
-        expect(protractor.ActionSequence).toBeDefined();
-        expect(protractor.Key.RETURN).toEqual('\uE006');
-    });
+        browser.get('index.html');
+        expect(browser.getCurrentUrl()).toMatch('#/form');
+
+        browser.driver.findElement(protractor.By.linkText('repeater')).click();
+        expect(browser.driver.getCurrentUrl()).toMatch('#/repeater');
+
+        browser.navigate().back();
+        expect(browser.driver.getCurrentUrl()).toMatch('#/form');
+      });
 
   it('should allow adding custom locators', function() {
     var findMenuItem = function() {
@@ -89,18 +80,19 @@ describe('protractor library', function() {
 
     browser.get('index.html');
     expect(element(by.menuItemWithName('.menu li', 'repeater')).isPresent());
-    expect(element(by.menuItemWithName('.menu li', 'repeater')).getText()).toEqual('repeater');
+    expect(element(by.menuItemWithName('.menu li', 'repeater')).getText()).
+        toEqual('repeater');
   });
 
   describe('helper functions', function() {
     it('should get the absolute URL', function() {
       browser.get('index.html');
       expect(browser.getLocationAbsUrl()).
-          toEqual('http://localhost:'+port+'/index.html#/form');
+          toMatch('index.html#/form');
 
       element(by.linkText('repeater')).click();
       expect(browser.getLocationAbsUrl()).
-          toEqual('http://localhost:'+port+'/index.html#/repeater');
+          toMatch('index.html#/repeater');
     });
 
     it('should navigate to another url with setLocation', function() {
@@ -109,7 +101,7 @@ describe('protractor library', function() {
       browser.setLocation('/repeater');
 
       expect(browser.getLocationAbsUrl()).
-        toEqual('http://localhost:' + port + '/index.html#/repeater');
+          toMatch('index.html#/repeater');
     });
   });
 });
